feat: Add comprehensive Neovim configuration with enhanced functionality
- Implemented various autocmds for improved user experience, including:
  - Highlighting on yank
  - Resizing splits on window resize
  - Automatic file change checks
  - Git commit message enhancements
  - Ensuring parent directories exist for new files
  - Help window positioning
  - Git and Fugitive buffer configurations

- Integrated completion setup using nvim-cmp and copilot, with custom mappings for snippets and sources.

- Established key mappings for efficient navigation, buffer management, and Git operations.

- Configured LSP with multiple language servers and diagnostic settings, including custom signs for diagnostics.

- Added plugin management with a focus on UI enhancements, coding tools, and Treesitter support.

- Set up user interface elements including lualine, nvim-tree, and notification handling.

- Enhanced Treesitter configuration with various language support and text object mappings.

- Implemented terminal management with toggleterm for better terminal integration.

- Created utility functions for buffer cycling and quickfix management.
