+++ home-manager/programs/neovim/init.lua
@@ -1,1330 +1,25 @@
 -- ====================================================================================
--- VIM OPTIONS
+-- LEADER KEY
 -- ====================================================================================
-vim.opt.compatible = false
-vim.opt.termsync = true
-vim.opt.hidden = true
-vim.opt.updatetime = 300
-vim.opt.mouse = ""
-vim.opt.inccommand = "nosplit"
-vim.opt.splitbelow = true
-vim.opt.splitright = true
-vim.opt.wrap = true
-vim.opt.textwidth = 0
-vim.opt.expandtab = true
-vim.opt.smartindent = true
-vim.opt.shiftwidth = 2
-vim.opt.softtabstop = 2
-vim.opt.tabstop = 2
-vim.opt.signcolumn = "yes"
-vim.opt.scrolloff = 10
-vim.opt.sidescrolloff = 10
-vim.opt.number = true
-vim.opt.relativenumber = true
-vim.opt.smoothscroll = true
-vim.opt.swapfile = false
-vim.opt.backup = true
-vim.opt.undofile = true
-vim.opt.undodir = vim.fn.stdpath("data") .. "undodir"
-vim.opt.hlsearch = false
-vim.opt.ignorecase = true
-vim.opt.incsearch = true
-vim.opt.ruler = true
-vim.opt.wildmenu = true
-vim.opt.autoread = true
-vim.opt.completeopt = { "menu", "menuone", "noselect" }
-vim.opt.colorcolumn = "80"
-vim.opt.backspace = { "indent", "eol", "start" }
-vim.opt.spelllang = { "en_us" }
-vim.opt.laststatus = 2
-vim.opt.cursorline = true
-vim.opt.grepprg = "rg --vimgrep --smart-case --follow"
--- Background will be set automatically based on system theme
-vim.opt.termguicolors = true
-vim.opt.shortmess:append("c")
-vim.opt.timeoutlen = 300
-vim.opt.winborder = "none"
-
-vim.hl.priorities.semantic_tokens = 10
-vim.g.fugitive_legacy_commands = 0
-
--- ====================================================================================
--- PLUGIN INSTALLATION
--- ====================================================================================
-vim.pack.add({
-	-- UI
-	{ src = "https://github.com/nvim-tree/nvim-web-devicons" },
-	{ src = "https://github.com/Mofiqul/dracula.nvim" },
-	{
-		src = "https://github.com/f-person/auto-dark-mode.nvim",
-		priority = 1000,
-	},
-	{ src = "https://github.com/folke/sidekick.nvim" },
-	{ src = "https://github.com/nvim-tree/nvim-tree.lua" },
-	{ src = "https://github.com/nvim-lualine/lualine.nvim" },
-	{ src = "https://github.com/stevearc/dressing.nvim" },
-	{ src = "https://github.com/rcarriga/nvim-notify" },
-	{ src = "https://github.com/christoomey/vim-tmux-navigator" },
-	{ src = "https://github.com/asiryk/auto-hlsearch.nvim" },
-	{ src = "https://github.com/famiu/bufdelete.nvim" },
-	{ src = "https://github.com/norcalli/nvim-colorizer.lua" },
-	{ src = "https://github.com/lewis6991/gitsigns.nvim" },
-	{ src = "https://github.com/akinsho/git-conflict.nvim" },
-
-	-- TELESCOPE
-	{ src = "https://github.com/nvim-lua/plenary.nvim" },
-	{ src = "https://github.com/nvim-telescope/telescope-github.nvim" },
-	{ src = "https://github.com/nvim-telescope/telescope.nvim" },
-
-	-- CODING
-	{ src = "https://github.com/rgroli/other.nvim" },
-	{ src = "https://github.com/danymat/neogen" },
-	{ src = "https://github.com/stevearc/conform.nvim" },
-	{ src = "https://github.com/zbirenbaum/copilot.lua" },
-	{ src = "https://github.com/rafamadriz/friendly-snippets" },
-	{ src = "https://github.com/giuxtaposition/blink-cmp-copilot" },
-	{
-		src = "https://github.com/saghen/blink.cmp",
-		version = vim.version.range("1.*"),
-	},
-
-	-- TPOPE
-	{ src = "https://github.com/tpope/vim-fugitive" },
-	{ src = "https://github.com/tpope/vim-rhubarb" },
-	{ src = "https://github.com/tpope/vim-abolish" },
-	{ src = "https://github.com/tpope/vim-repeat" },
-	{ src = "https://github.com/tpope/vim-eunuch" },
-	{ src = "https://github.com/tpope/vim-sleuth" },
-	{ src = "https://github.com/tpope/vim-speeddating" },
-
-	-- TREESITTER
-	{ src = "https://github.com/nvim-treesitter/nvim-treesitter" },
-	{ src = "https://github.com/lukas-reineke/indent-blankline.nvim" },
-	{ src = "https://github.com/wansmer/treesj" },
-	{ src = "https://github.com/windwp/nvim-autopairs" },
-	{ src = "https://github.com/windwp/nvim-ts-autotag" },
-	{ src = "https://github.com/kylechui/nvim-surround" },
-	{ src = "https://github.com/folke/todo-comments.nvim" },
-	{ src = "https://github.com/nvim-treesitter/nvim-treesitter-context" },
-	{ src = "https://github.com/nvim-treesitter/nvim-treesitter-textobjects" },
-	{ src = "https://github.com/RRethy/nvim-treesitter-endwise" },
-})
-
--- ====================================================================================
--- KEYMAPS
--- ====================================================================================
-local opts = { noremap = true, silent = true }
-local keymap = vim.keymap.set
-
--- @keymap <Space>: Set as leader key
---Remap space as leader key
-keymap("", "<Space>", "<Nop>", opts)
 vim.g.mapleader = " "
 vim.g.maplocalleader = " "
 
 -- ====================================================================================
--- QUICKFIX LISTS
+-- Set path to modules
 -- ====================================================================================
--- @keymap <leader>co: Open quickfix list
-keymap("n", "<leader>co", ":copen<CR>", opts)
--- @keymap <leader>cc: Close quickfix list
-keymap("n", "<leader>cc", ":cclose<CR>", opts)
+package.path = package.path .. ";" .. vim.fn.stdpath("config") .. "/lua/?.lua"
 
 -- ====================================================================================
--- BUFFER AND FILE OPERATIONS
+-- MODULES
 -- ====================================================================================
--- Helper function to cycle buffers (includes unlisted buffers like nvim-tree)
-local function cycle_buffer(direction)
-	local current_buf = vim.api.nvim_get_current_buf()
-
-	-- Get all loaded buffers (this includes buffers in windows and loaded but not visible)
-	local buffer_set = {}
-	local buffers = {}
-
-	-- First, get buffers from windows (includes nvim-tree)
-	for _, win in ipairs(vim.api.nvim_list_wins()) do
-		local buf = vim.api.nvim_win_get_buf(win)
-		if vim.api.nvim_buf_is_valid(buf) and vim.api.nvim_buf_is_loaded(buf) and not buffer_set[buf] then
-			buffer_set[buf] = true
-			table.insert(buffers, buf)
-		end
-	end
-
-	-- Then add all other loaded buffers (files that aren't in windows)
-	for _, buf in ipairs(vim.api.nvim_list_bufs()) do
-		if vim.api.nvim_buf_is_valid(buf) and vim.api.nvim_buf_is_loaded(buf) and not buffer_set[buf] then
-			buffer_set[buf] = true
-			table.insert(buffers, buf)
-		end
-	end
-
-	-- Ensure current buffer is in the list
-	if not buffer_set[current_buf] then
-		table.insert(buffers, current_buf)
-		buffer_set[current_buf] = true
-	end
-
-	-- Need at least 2 buffers to cycle
-	if #buffers < 2 then
-		return
-	end
-
-	-- Find current buffer index
-	local current_idx = nil
-	for i, buf in ipairs(buffers) do
-		if buf == current_buf then
-			current_idx = i
-			break
-		end
-	end
-
-	-- This should never happen now, but safety check
-	if not current_idx then
-		current_idx = 1
-	end
-
-	-- Calculate next/previous index with wrapping
-	local next_idx
-	if direction == "next" then
-		next_idx = current_idx + 1
-		if next_idx > #buffers then
-			next_idx = 1
-		end
-	else -- prev
-		next_idx = current_idx - 1
-		if next_idx < 1 then
-			next_idx = #buffers
-		end
-	end
-
-	-- Get target buffer
-	local target_buf = buffers[next_idx]
-
-	-- Safety check
-	if not vim.api.nvim_buf_is_valid(target_buf) then
-		return
-	end
-
-	-- Find window containing target buffer
-	local target_win = nil
-	for _, win in ipairs(vim.api.nvim_list_wins()) do
-		if vim.api.nvim_win_is_valid(win) and vim.api.nvim_win_get_buf(win) == target_buf then
-			target_win = win
-			break
-		end
-	end
-
-	if target_win then
-		-- Buffer is in a window, switch to that window
-		vim.api.nvim_set_current_win(target_win)
-	else
-		-- Buffer not in any window, switch to it in current window
-		vim.api.nvim_set_current_buf(target_buf)
-	end
-end
-
--- @keymap <leader><Tab>: Cycle to next buffer
-vim.keymap.set("n", "<leader><tab>", function()
-	cycle_buffer("next")
-end, { noremap = true, silent = true })
-
--- @keymap <leader><S-Tab>: Cycle to previous buffer
-vim.keymap.set("n", "<leader><s-tab>", function()
-	cycle_buffer("prev")
-end, { noremap = true, silent = true })
-
--- Alternative: ]b and [b for buffer navigation (vim convention)
--- @keymap ]b: Cycle to next buffer
-keymap("n", "]b", function()
-	cycle_buffer("next")
-end, opts)
-
--- @keymap [b: Cycle to previous buffer
-keymap("n", "[b", function()
-	cycle_buffer("prev")
-end, opts)
--- @keymap <leader>q: Close current buffer
-keymap("n", "<leader>q", ":Bdelete<CR>", opts)
--- @keymap <leader>bad: Wipe all buffers
-keymap("n", "<leader>bad", ":%bwipeout!<cr>:intro<cr>", opts)
--- @keymap <leader>w: Write file
-keymap("n", "<leader>w", ":write<CR>", opts)
--- @keymap <leader>r: Reload Neovim configuration
-keymap("n", "<leader>r", function()
-	vim.cmd("source $MYVIMRC")
-end, opts)
--- @keymap <leader>h: Show help (all keymaps and commands)
-keymap("n", "<leader>h", ":Help<CR>", opts)
--- @keymap ZZ: Save all buffers and quit
-keymap("n", "ZZ", ":wa<CR>:q<CR>", opts)
-
--- ====================================================================================
--- TERMINAL
--- ====================================================================================
--- Terminal toggle state
-local terminal_bufnr = nil
-local terminal_winid = nil
-
--- Toggle terminal function
-local function toggle_terminal()
-	-- Check if terminal window is currently visible
-	if terminal_winid and vim.api.nvim_win_is_valid(terminal_winid) then
-		-- Terminal is visible, hide it (but keep the buffer)
-		vim.api.nvim_win_hide(terminal_winid)
-		terminal_winid = nil
-		return
-	end
-
-	-- Check if terminal buffer exists
-	if terminal_bufnr and vim.api.nvim_buf_is_valid(terminal_bufnr) then
-		-- Check if buffer is already displayed in any window
-		local existing_winid = nil
-		for _, winid in ipairs(vim.api.nvim_list_wins()) do
-			if vim.api.nvim_win_get_buf(winid) == terminal_bufnr then
-				existing_winid = winid
-				break
-			end
-		end
-
-		if existing_winid then
-			-- Buffer is already displayed, focus that window
-			vim.api.nvim_set_current_win(existing_winid)
-			terminal_winid = existing_winid
-		else
-			-- Terminal buffer exists but window is hidden, show it
-			vim.cmd("10split")
-			vim.api.nvim_set_current_buf(terminal_bufnr)
-			terminal_winid = vim.api.nvim_get_current_win()
-		end
-		return
-	end
-
-	-- Terminal doesn't exist, create it
-	vim.cmd("10split | terminal")
-	terminal_bufnr = vim.api.nvim_get_current_buf()
-	terminal_winid = vim.api.nvim_get_current_win()
-end
-
--- @keymap <leader>j: Toggle terminal (show/hide)
-keymap("n", "<leader>j", toggle_terminal, opts)
-
--- ====================================================================================
--- NAVIGATION
--- ====================================================================================
--- @keymap n: Next search result and center screen
-keymap("n", "n", "nzzzv", opts)
--- @keymap N: Previous search result and center screen
-keymap("n", "N", "Nzzzv", opts)
--- @keymap <C-u>: Scroll up and center screen
-keymap("n", "<C-u>", "<C-u>zz", opts)
--- @keymap <C-d>: Scroll down and center screen
-keymap("n", "<C-d>", "<C-d>zz", opts)
--- @keymap <C-o>: Jump to older position and center screen
-keymap("n", "<C-o>", "<C-o>zz", opts)
--- @keymap <C-i>: Jump to newer position and center screen
-keymap("n", "<C-i>", "<C-i>zz", opts)
-
--- ====================================================================================
--- CLIPBOARD OPERATIONS
--- ====================================================================================
--- @keymap <leader>y: Yank to system clipboard
--- system clipboard integration
-keymap({ "n", "v" }, "<leader>y", '"+y', opts)
-
--- @keymap <leader>py: Copy current file path to clipboard
--- copy the current file path
-keymap("n", "<leader>py", ':let @" = expand("%:p")<CR>', opts)
-
--- @keymap <leader>d: Delete to blackhole register
--- delete to blackhole
-keymap({ "n", "v" }, "<leader>d", '"_d', opts)
-
--- ====================================================================================
--- GIT OPERATIONS
--- ====================================================================================
--- @keymap <leader>gs: Open Git status in new tab
-keymap("n", "<leader>gs", ":tab Git<cr>", opts)
--- @keymap <F9>: Open Git mergetool in new tab
-keymap("n", "<F9>", ":tab Git mergetool<cr>", opts)
--- @keymap <leader>gd: Preview hunk inline
-keymap("n", "<leader>gd", ":Gitsign preview_hunk_inline<cr>", opts)
-
--- ====================================================================================
--- VISUAL MODE OPERATIONS
--- ====================================================================================
--- @keymap <: Decrease indent (visual mode)
-keymap("v", "<", "<gv", opts)
--- @keymap >: Increase indent (visual mode)
-keymap("v", ">", ">gv", opts)
-
--- @keymap p: Paste without replacing clipboard (visual mode)
--- If I visually select words and paste from clipboard, don't replace my
--- clipboard with the selected word, instead keep my old word in the
--- clipboard
-keymap("v", "p", '"_dP', opts)
-
--- @keymap K: Move selected text up (visual mode)
-keymap("x", "K", ":move '<-2<CR>gv-gv", opts)
--- @keymap J: Move selected text down (visual mode)
-keymap("x", "J", ":move '>+1<CR>gv-gv", opts)
-
--- ====================================================================================
--- INSERT MODE OPERATIONS
--- ====================================================================================
--- @keymap -/_/,/./;/:/!/?: Add undo breakpoints after punctuation
--- in insert mode, adds new undo points after more some chars:
-for _, lhs in ipairs({ "-", "_", ",", ".", ";", ":", "/", "!", "?" }) do
-	keymap("i", lhs, lhs .. "<c-g>u", opts)
-end
-
--- @keymap jj: Exit insert mode
--- exit insert mode with jj
-keymap("i", "jj", "<Esc>", opts)
-
--- ====================================================================================
--- UI CONFIGURATION
--- ====================================================================================
-
-vim.api.nvim_set_hl(0, "StatusLine", { reverse = false })
-vim.api.nvim_set_hl(0, "StatusLineNC", { reverse = false })
-
-require("dressing").setup({ input = { insert_only = true } })
-
-local notify = require("notify")
-notify.setup({
-	render = "compact",
-	stages = "static",
-})
-vim.notify = notify
-
-local section_b = { "branch", "diff", { "diagnostics", sources = { "nvim_workspace_diagnostic" } } }
-local section_c = { "%=", { "filename", file_status = false, path = 1 } }
-local lualine_config = {
-	options = {
-		theme = function()
-			-- Dynamically determine theme based on background
-			return vim.o.background == "dark" and "dracula" or "auto"
-		end,
-		component_separators = "",
-		section_separators = "",
-	},
-	sections = {
-		lualine_b = section_b,
-		lualine_c = section_c,
-	},
-	inactive_sections = {
-		lualine_c = section_c,
-		lualine_x = { "location" },
-	},
-}
-require("lualine").setup(lualine_config)
-
--- Auto theme switching based on system appearance
-require("auto-dark-mode").setup({
-	update_interval = 1000, -- Check for theme changes every second
-	set_dark_mode = function()
-		vim.api.nvim_set_option_value("background", "dark", {})
-		vim.cmd("colorscheme dracula")
-		-- Refresh lualine to update theme
-		if package.loaded["lualine"] then
-			require("lualine").setup(lualine_config)
-		end
-	end,
-	set_light_mode = function()
-		vim.api.nvim_set_option_value("background", "light", {})
-		vim.cmd("colorscheme default")
-		-- Refresh lualine to update theme
-		if package.loaded["lualine"] then
-			require("lualine").setup(lualine_config)
-		end
-	end,
-})
-
-require("auto-hlsearch").setup({})
-require("gitsigns").setup({})
-require("git-conflict").setup({})
-
--- ====================================================================================
--- DIAGNOSTICS
--- ====================================================================================
--- setup diagnostics
-vim.diagnostic.config({
-	underline = true,
-	update_in_insert = false,
-	severity_sort = true,
-})
-
--- @keymap <leader>xx: Open diagnostics in quickfix
-keymap("n", "<leader>xx", vim.diagnostic.setqflist, opts)
-
--- set up diagnostic signs
-local severity = vim.diagnostic.severity
-local signs = {
-	[severity.ERROR] = "",
-	[severity.WARN] = "",
-	[severity.INFO] = "",
-	[severity.HINT] = "󰌶",
-}
-vim.diagnostic.config({
-	signs = {
-		text = signs,
-	},
-})
-
--- ====================================================================================
--- CODE NAVIGATION
--- ====================================================================================
-require("other-nvim").setup({ mappings = { "golang" } })
--- @keymap <leader>oo: Navigate to related file (other.nvim)
-keymap("n", "<leader>oo", ":Other<cr>", opts)
--- @keymap <leader>ov: Navigate to related file in vertical split
-keymap("n", "<leader>ov", ":OtherVSplit<cr>", opts)
--- @keymap <leader>os: Navigate to related file in horizontal split
-keymap("n", "<leader>os", ":OtherSplit<cr>", opts)
-
--- @keymap gco: Generate code annotation (neogen)
-require("neogen").setup()
-keymap("n", "gco", ":Neogen<cr>", opts)
-
--- ====================================================================================
--- COMPLETION AND COPILOT
--- ====================================================================================
-require("copilot").setup({
-	suggestion = { enabled = false },
-	panel = { enabled = false },
-})
-
----@module 'blink.cmp'
----@type blink.cmp.Config
-local cmp = require("blink.cmp")
-cmp.setup({
-	appearance = {
-		nerd_font_variant = "mono",
-		kind_icons = {
-			Array = "",
-			Boolean = "",
-			Class = "",
-			Color = "",
-			Constant = "",
-			Constructor = "",
-			Copilot = "",
-			Enum = "",
-			EnumMember = "",
-			Event = "",
-			Field = "",
-			File = "",
-			Folder = "󰉋",
-			Function = "",
-			Interface = "",
-			Key = "",
-			Keyword = "",
-			Method = "",
-			Module = "",
-			Namespace = "",
-			Null = "󰟢",
-			Number = "",
-			Object = "",
-			Operator = "",
-			Package = "",
-			Property = "",
-			Reference = "",
-			Snippet = "",
-			String = "",
-			Struct = "",
-			Text = "",
-			TypeParameter = "",
-			Unit = "",
-			Value = "",
-			Variable = "",
-		},
-	},
-	cmdline = {
-		sources = function()
-			local type = vim.fn.getcmdtype()
-			if type == "/" or type == "?" then
-				return { "buffer" }
-			end
-			if type == ":" then
-				return { "cmdline" }
-			end
-			return {}
-		end,
-		completion = {
-			menu = {
-				draw = {
-					columns = {
-						{ "kind_icon", "label", gap = 1 },
-						{ "kind" },
-					},
-				},
-			},
-		},
-	},
-	sources = {
-		default = { "lsp", "path", "snippets", "buffer", "copilot" },
-		providers = {
-			lsp = {
-				min_keyword_length = 0,
-				score_offset = 0,
-			},
-			path = {
-				min_keyword_length = 0,
-			},
-			snippets = {
-				min_keyword_length = 2,
-			},
-			buffer = {
-				min_keyword_length = 5,
-				max_items = 5,
-			},
-			copilot = {
-				name = "copilot",
-				module = "blink-cmp-copilot",
-				score_offset = 1000,
-				min_keyword_length = 0,
-				async = true,
-				override = {
-					-- copilot complete on space, new line, etc as well...
-					get_trigger_characters = function(self)
-						local trigger_characters = self:get_trigger_characters()
-						vim.list_extend(trigger_characters, { "\n", "\t", " " })
-						return trigger_characters
-					end,
-				},
-				transform_items = function(_, items)
-					local CompletionItemKind = require("blink.cmp.types").CompletionItemKind
-					local kind_idx = #CompletionItemKind + 1
-					CompletionItemKind[kind_idx] = "Copilot"
-					for _, item in ipairs(items) do
-						item.kind = kind_idx
-					end
-					return items
-				end,
-			},
-		},
-	},
-	completion = {
-		accept = { auto_brackets = { enabled = true } },
-		keyword = {
-			range = "full",
-		},
-		menu = {
-			draw = {
-				columns = {
-					{ "kind_icon", "label", gap = 1 },
-					{ "kind" },
-				},
-			},
-		},
-	},
-})
-
--- @keymap <Tab>: Accept copilot/completion (selects first if none selected)
--- Use Tab to accept copilot/completions
-keymap("i", "<Tab>", function()
-	-- Check if completion menu is visible (works for blink.cmp and built-in)
-	if vim.fn.pumvisible() == 1 then
-		-- Since noselect is enabled, first select the first item, then accept
-		-- This ensures we accept the top suggestion (or navigate to copilot if needed)
-		local current_selection = vim.fn.complete_info({ "selected" }).selected
-		if current_selection == -1 then
-			-- Nothing selected, select first item
-			return vim.api.nvim_replace_termcodes("<C-n><C-y>", true, false, true)
-		else
-			-- Something is selected, just accept it
-			return vim.api.nvim_replace_termcodes("<C-y>", true, false, true)
-		end
-	end
-	-- No completion menu visible, insert tab normally
-	return vim.api.nvim_replace_termcodes("<Tab>", true, false, true)
-end, { expr = true, noremap = true, silent = true })
-
--- ====================================================================================
--- UTILITY FUNCTIONS
--- ====================================================================================
-local function copen()
-	if vim.fn.getqflist({ size = 0 }).size > 1 then
-		vim.cmd("copen")
-	else
-		vim.cmd("cclose")
-	end
-end
-
-local function cclear()
-	vim.fn.setqflist({}, "r")
-end
-
--- @command Finder: Open current file directory in Finder/file explorer
--- Opens the directory of the current file in Finder/file explorer.
-vim.api.nvim_create_user_command("Finder", "!open %:h", {})
-
--- @command ThemeRefresh: Refresh theme based on system appearance
--- Refreshes the colorscheme based on current system theme (dark/light).
-vim.api.nvim_create_user_command("ThemeRefresh", function()
-	require("auto-dark-mode").update()
-end, { desc = "Refresh theme based on system appearance" })
-
--- ====================================================================================
--- AUTOCMDS
--- ====================================================================================
--- Check for file changes on BufEnter, CursorHold, CursorHoldI, and FocusGained
-vim.api.nvim_create_autocmd({
-	"BufEnter",
-	"CursorHold",
-	"CursorHoldI",
-	"FocusGained",
-}, {
-	pattern = "*",
-	command = "if mode() != 'c' | checktime | endif",
-})
-
--- Resize splits if window got resized
-vim.api.nvim_create_autocmd({ "VimResized" }, {
-	callback = function()
-		vim.cmd("tabdo wincmd =")
-	end,
-})
-
--- Start in insert mode when opening git commit messages
-vim.api.nvim_create_autocmd("FileType", {
-	pattern = "gitcommit",
-	command = "startinsert",
-})
-
--- Ensure the parent folder exists when creating new files (for LSP context)
-vim.api.nvim_create_autocmd("BufNewFile", {
-	pattern = "*",
-	callback = function()
-		local dir = vim.fn.expand("<afile>:p:h")
-		if vim.fn.isdirectory(dir) == 0 then
-			vim.fn.mkdir(dir, "p")
-			vim.cmd([[ :e % ]])
-		end
-	end,
-})
-
--- Highlight on yank (briefly highlight yanked text)
--- See `:help vim.highlight.on_yank()`
-vim.api.nvim_create_autocmd("TextYankPost", {
-	pattern = "*",
-	callback = function()
-		vim.hl.on_yank()
-	end,
-})
-
--- Open help window in a vertical split to the right
-vim.api.nvim_create_autocmd("BufWinEnter", {
-	pattern = { "*.txt" },
-	callback = function()
-		if vim.o.filetype == "help" then
-			vim.cmd.wincmd("L")
-		end
-	end,
-})
-
--- Add keymaps for git filetype buffers
-vim.api.nvim_create_autocmd("FileType", {
-	pattern = "git",
-	callback = function()
-		local bufnr = vim.api.nvim_get_current_buf()
-		local buf_opts = { noremap = true, silent = true, buffer = bufnr }
-		-- @keymap gq: Close git buffer
-		keymap("n", "gq", ":silent! close<cr>", buf_opts)
-	end,
-})
-
--- Configure fugitive buffers with git keymaps (push, pull, commit, etc.)
-vim.api.nvim_create_autocmd("FileType", {
-	pattern = "fugitive",
-	callback = function()
-		local bufnr = vim.api.nvim_get_current_buf()
-
-		local function async_git(args, success_msg, error_msg)
-			vim.system({ "git", unpack(args) }, {}, function(obj)
-				vim.schedule(function()
-					if obj.code == 0 then
-						vim.notify(success_msg, vim.log.levels.INFO)
-					else
-						vim.notify(error_msg, vim.log.levels.ERROR)
-					end
-				end)
-			end)
-		end
-
-		vim.cmd("normal )k=")
-
-		local buf_opts = { noremap = true, silent = true, buffer = bufnr }
-		-- @keymap gp: Git push (fugitive buffer)
-		keymap("n", "gp", function()
-			async_git({ "push", "--quiet" }, "Pushed!", "Push failed!")
-			vim.cmd("silent! close")
-		end, buf_opts)
-
-		-- @keymap gP: Git pull with rebase (fugitive buffer)
-		keymap("n", "gP", function()
-			async_git({ "pull", "--rebase" }, "Pulled!", "Pull failed!")
-			vim.cmd("silent! close")
-		end, buf_opts)
-
-		-- @keymap go: Git push and open PR (fugitive buffer)
-		keymap("n", "go", function()
-			async_git({ "ppr" }, "Pushed and opened PR URL!", "Failed to push or open PR")
-			vim.cmd("silent! close")
-		end, buf_opts)
-
-		-- @keymap cc: Git commit with sign-off (fugitive buffer)
-		-- @keymap gq: Close fugitive buffer
-		keymap("n", "cc", ":silent! Git commit -s<cr>", buf_opts)
-		keymap("n", "gq", ":silent! close<cr>", buf_opts)
-	end,
-})
-
--- Enable spell check and set textwidth for git commit messages
-vim.api.nvim_create_autocmd("FileType", {
-	pattern = "gitcommit",
-	callback = function()
-		vim.opt_local.spell = true
-		vim.opt_local.textwidth = 72
-	end,
-})
-
--- Add <leader>q keymap to close quickfix and help buffers
-vim.api.nvim_create_autocmd("FileType", {
-	pattern = { "qf", "help" },
-	callback = function()
-		-- @keymap <leader>q: Close quickfix/help buffer
-		keymap("n", "<leader>q", ":bdelete<CR>", {
-			buffer = vim.api.nvim_get_current_buf(),
-			noremap = true,
-			silent = true,
-		})
-	end,
-})
-
--- Track terminal buffer and window, and set up terminal mode keymap
-vim.api.nvim_create_autocmd("TermOpen", {
-	callback = function()
-		local bufnr = vim.api.nvim_get_current_buf()
-		local winid = vim.api.nvim_get_current_win()
-
-		-- Update terminal tracking if this is our managed terminal
-		if terminal_bufnr == nil or terminal_bufnr == bufnr then
-			terminal_bufnr = bufnr
-			terminal_winid = winid
-		end
-
-		-- @keymap <leader>j: Exit terminal mode and hide terminal
-		keymap("t", "<leader>j", function()
-			-- Exit terminal mode properly
-			vim.api.nvim_feedkeys(vim.api.nvim_replace_termcodes("<C-\\><C-N>", true, false, true), "n", false)
-			-- Hide the terminal window but keep the buffer
-			vim.schedule(function()
-				if vim.api.nvim_win_is_valid(terminal_winid) then
-					vim.api.nvim_win_hide(terminal_winid)
-					terminal_winid = nil
-				end
-			end)
-		end, {
-			buffer = bufnr,
-			noremap = true,
-			silent = true,
-		})
-
-		-- @keymap <Esc>: Exit terminal mode (keeps terminal visible)
-		-- Standard Neovim way: <C-\><C-n> also works
-		keymap("t", "<Esc>", vim.api.nvim_replace_termcodes("<C-\\><C-N>", true, false, true), {
-			buffer = bufnr,
-			noremap = true,
-			silent = true,
-		})
-
-		-- @keymap <Tab>: Trigger completion in terminal mode
-		-- Note: Terminal completion works differently - you may need to use <C-x><C-f> for file completion
-		-- or configure your shell's completion (e.g., fish, zsh with completion plugins)
-	end,
-})
-
--- Clean up terminal tracking when terminal buffer is closed
-vim.api.nvim_create_autocmd("BufDelete", {
-	callback = function(args)
-		if args.buf == terminal_bufnr then
-			terminal_bufnr = nil
-			terminal_winid = nil
-		end
-	end,
-})
-
--- Enable spell check and set textwidth for markdown files
-vim.api.nvim_create_autocmd("FileType", {
-	pattern = "markdown",
-	callback = function()
-		vim.opt_local.spell = true
-		vim.opt_local.textwidth = 80
-		vim.opt_local.formatoptions:remove("ct")
-	end,
-})
-
--- ====================================================================================
--- TREESITTER AND SYNTAX
--- ====================================================================================
--- syntax, indentation, treesitter, etc.
-require("ibl").setup({
-	indent = { char = "│" },
-	exclude = { filetypes = { "help" } },
-	scope = { enabled = false },
-})
-
-require("nvim-autopairs").setup()
-
-require("nvim-ts-autotag").setup()
-require("nvim-surround").setup()
-require("todo-comments").setup()
-
-require("treesitter-context").setup()
-
--- ====================================================================================
--- TREESITTER CONFIGURATION
--- ====================================================================================
----@diagnostic disable-next-line: missing-fields
-require("nvim-treesitter.configs").setup({
-	ensure_installed = {
-		"arduino",
-		"awk",
-		"bash",
-		"cpp",
-		"css",
-		"csv",
-		"diff",
-		"dockerfile",
-		"fish",
-		"git_config",
-		"git_rebase",
-		"gitattributes",
-		"gitcommit",
-		"gitignore",
-		"go",
-		"gomod",
-		"gosum",
-		"gowork",
-		"graphql",
-		"hcl",
-		"html",
-		"http",
-		"http",
-		"ini",
-		"javascript",
-		"jq",
-		"json",
-		"lua",
-		"make",
-		"markdown",
-		"markdown_inline",
-		"mermaid",
-		"python",
-		"query",
-		"regex",
-		"ruby",
-		"scss",
-		"sql",
-		"ssh_config",
-		"templ",
-		"terraform",
-		"toml",
-		"vhs",
-		"vim",
-		"vimdoc",
-		"yaml",
-		"zig",
-	},
-	highlight = {
-		enable = true,
-	},
-	indent = {
-		enable = true,
-	},
-	endwise = {
-		enable = true,
-	},
-	autopairs = {
-		enable = true,
-	},
-	context_commentstring = {
-		enable = true,
-		enable_autocmd = false,
-	},
-	incremental_selection = {
-		enable = true,
-		keymaps = {
-			init_selection = "<C-space>",
-			node_incremental = "<C-space>",
-			node_decremental = "<bs>",
-			scope_incremental = "<noop>",
-		},
-	},
-	auto_install = true,
-	textobjects = {
-		enable = true,
-		lookahead = true,
-		swap = {
-			enable = true,
-			swap_next = {
-				-- @keymap <leader>a: Swap next parameter (treesitter)
-				["<leader>a"] = "@parameter.inner",
-			},
-			swap_previous = {
-				-- @keymap <leader>A: Swap previous parameter (treesitter)
-				["<leader>A"] = "@parameter.inner",
-			},
-		},
-		move = {
-			enable = true,
-			set_jumps = true,
-			goto_next_start = {
-				-- @keymap ]f: Go to next function start (treesitter)
-				-- @keymap ]c: Go to next class start (treesitter)
-				-- @keymap ]a: Go to next parameter start (treesitter)
-				["]f"] = "@function.inner",
-				["]c"] = "@class.inner",
-				["]a"] = "@parameter.inner",
-			},
-			goto_next_end = {
-				-- @keymap ]F: Go to next function end (treesitter)
-				-- @keymap ]C: Go to next class end (treesitter)
-				-- @keymap ]A: Go to next parameter end (treesitter)
-				["]F"] = "@function.inner",
-				["]C"] = "@class.inner",
-				["]A"] = "@parameter.inner",
-			},
-			goto_previous_start = {
-				-- @keymap [f: Go to previous function start (treesitter)
-				-- @keymap [c: Go to previous class start (treesitter)
-				-- @keymap [a: Go to previous parameter start (treesitter)
-				["[f"] = "@function.inner",
-				["[c"] = "@class.inner",
-				["[a"] = "@parameter.inner",
-			},
-			goto_previous_end = {
-				-- @keymap [F: Go to previous function end (treesitter)
-				-- @keymap [C: Go to previous class end (treesitter)
-				-- @keymap [A: Go to previous parameter end (treesitter)
-				["[F"] = "@function.inner",
-				["[C"] = "@class.inner",
-				["[A"] = "@parameter.inner",
-			},
-		},
-		select = {
-			enable = true,
-			keymaps = {
-				-- @keymap af: Select function outer (treesitter)
-				-- @keymap if: Select function inner (treesitter)
-				-- @keymap ac: Select conditional outer (treesitter)
-				-- @keymap ic: Select conditional inner (treesitter)
-				-- @keymap aa: Select parameter outer (treesitter)
-				-- @keymap ia: Select parameter inner (treesitter)
-				-- @keymap av: Select variable outer (treesitter)
-				-- @keymap iv: Select variable inner (treesitter)
-				["af"] = "@function.outer",
-				["if"] = "@function.inner",
-
-				["ac"] = "@conditional.outer",
-				["ic"] = "@conditional.inner",
-
-				["aa"] = "@parameter.outer",
-				["ia"] = "@parameter.inner",
-
-				["av"] = "@variable.outer",
-				["iv"] = "@variable.inner",
-			},
-		},
-	},
-})
-
--- @keymap <leader>st: Toggle treesitter split/join
-local treesj = require("treesj")
-treesj.setup({ use_default_keymaps = false })
-keymap("n", "<leader>st", treesj.toggle, opts)
-
--- ====================================================================================
--- TELESCOPE CONFIGURATION
--- ====================================================================================
-local telescope = require("telescope")
-telescope.setup({
-	defaults = {
-		pickers = {
-			find_files = {
-				theme = "ivy",
-			},
-		},
-		prompt_prefix = "   ",
-		selection_caret = " ❯ ",
-		entry_prefix = "   ",
-		multi_icon = "+ ",
-		path_display = { "filename_first" },
-		vimgrep_arguments = {
-			"rg",
-			"--color=never",
-			"--no-heading",
-			"--with-filename",
-			"--line-number",
-			"--column",
-			"--smart-case",
-			"--sort=path",
-		},
-	},
-})
-telescope.load_extension("gh")
-
-local function ivy(iopts)
-	return require("telescope.themes").get_ivy(iopts)
-end
-
-local builtin = require("telescope.builtin")
--- @keymap <C-p>: Find files (Telescope)
-vim.keymap.set("n", "<C-p>", function()
-	builtin.find_files(ivy({
-		find_command = {
-			"fd",
-			"--type",
-			"f",
-			"--strip-cwd-prefix",
-			"--hidden",
-		},
-	}))
-end, opts)
-
--- @keymap <leader>of: Open old files (Telescope)
-keymap("n", "<leader>of", function()
-	builtin.oldfiles(ivy({
-		only_cwd = true,
-	}))
-end, opts)
-
--- @keymap <leader>lg: Live grep (Telescope)
-keymap("n", "<leader>lg", function()
-	builtin.live_grep(ivy())
-end, opts)
-
--- @keymap <leader>fb: Find buffers (Telescope)
-keymap("n", "<leader>fb", function()
-	builtin.buffers(ivy())
-end, opts)
-
--- @keymap <leader>fh: Find help tags (Telescope)
-keymap("n", "<leader>fh", function()
-	builtin.help_tags(ivy())
-end, opts)
-
--- @keymap <leader>fc: Find commands (Telescope)
-keymap("n", "<leader>fc", function()
-	builtin.commands(ivy())
-end, opts)
-
--- @keymap <leader>fr: Resume last Telescope search
-keymap("n", "<leader>fr", function()
-	builtin.resume(ivy())
-end, opts)
-
--- @keymap <leader>fq: Find in quickfix (Telescope)
-keymap("n", "<leader>fq", function()
-	builtin.quickfix(ivy())
-end, opts)
-
--- @keymap <leader>/: Fuzzy find in current buffer (Telescope)
-keymap("n", "<leader>/", function()
-	builtin.current_buffer_fuzzy_find(ivy())
-end, opts)
-
--- @keymap <leader>ghi: Find GitHub issues (Telescope)
-keymap("n", "<leader>ghi", function()
-	telescope.extensions.gh.issues(ivy())
-end, opts)
-
--- ====================================================================================
--- FILE TREE
--- ====================================================================================
--- @keymap <leader>b: Toggle file tree (nvim-tree)
-require("nvim-tree").setup({
-	view = {
-		width = 30,
-	},
-	actions = {
-		open_file = {
-			quit_on_open = false,
-		},
-	},
-	on_attach = function(bufnr)
-		local api = require("nvim-tree.api")
-
-		local function opts(desc)
-			return { desc = "nvim-tree: " .. desc, buffer = bufnr, noremap = true, silent = true, nowait = true }
-		end
-
-		-- Default mappings. You can override them in the mappings table below
-		-- @keymap l: Open file/folder (nvim-tree)
-		vim.keymap.set("n", "l", api.node.open.edit, opts("Open"))
-		-- @keymap <CR>: Open file/folder (nvim-tree)
-		vim.keymap.set("n", "<CR>", api.node.open.edit, opts("Open"))
-		-- @keymap o: Open file/folder (nvim-tree)
-		vim.keymap.set("n", "o", api.node.open.edit, opts("Open"))
-		-- @keymap h: Close folder (nvim-tree)
-		vim.keymap.set("n", "h", api.node.navigate.parent_close, opts("Close Directory"))
-		-- @keymap v: Open in vertical split (nvim-tree)
-		vim.keymap.set("n", "v", api.node.open.vertical, opts("Open: Vertical Split"))
-		-- @keymap s: Open in horizontal split (nvim-tree)
-		vim.keymap.set("n", "s", api.node.open.horizontal, opts("Open: Horizontal Split"))
-		-- @keymap t: Open in new tab (nvim-tree)
-		vim.keymap.set("n", "t", api.node.open.tab, opts("Open: New Tab"))
-		-- @keymap i: Open in horizontal split (nvim-tree)
-		vim.keymap.set("n", "i", api.node.open.horizontal, opts("Open: Horizontal Split"))
-		-- @keymap a: Add file/folder (nvim-tree)
-		vim.keymap.set("n", "a", api.fs.create, opts("Create"))
-		-- @keymap d: Delete file/folder (nvim-tree)
-		vim.keymap.set("n", "d", api.fs.remove, opts("Delete"))
-		-- @keymap r: Rename file/folder (nvim-tree)
-		vim.keymap.set("n", "r", api.fs.rename, opts("Rename"))
-		-- @keymap y: Copy absolute path (nvim-tree)
-		vim.keymap.set("n", "y", api.fs.copy.absolute_path, opts("Copy Absolute Path"))
-		-- @keymap Y: Copy relative path (nvim-tree)
-		vim.keymap.set("n", "Y", api.fs.copy.relative_path, opts("Copy Relative Path"))
-		-- @keymap c: Change directory (nvim-tree)
-		vim.keymap.set("n", "c", api.tree.change_root_to_node, opts("CD"))
-		-- @keymap u: Go to parent directory (nvim-tree)
-		vim.keymap.set("n", "u", api.tree.change_root_to_parent, opts("Up"))
-		-- @keymap q: Close tree (nvim-tree)
-		vim.keymap.set("n", "q", api.tree.close, opts("Close"))
-		-- @keymap R: Refresh tree (nvim-tree)
-		vim.keymap.set("n", "R", api.tree.reload, opts("Refresh"))
-		-- @keymap ?: Toggle help (nvim-tree)
-		vim.keymap.set("n", "?", api.tree.toggle_help, opts("Help"))
-		-- @keymap .: Run file command (nvim-tree)
-		vim.keymap.set("n", ".", api.node.run.cmd, opts("Run Command"))
-		-- @keymap <C-r>: Rename file/folder (nvim-tree)
-		vim.keymap.set("n", "<C-r>", api.fs.rename_basename, opts("Rename: Basename"))
-		-- @keymap <C-x>: Cut file/folder (nvim-tree)
-		vim.keymap.set("n", "<C-x>", api.fs.cut, opts("Cut"))
-		-- @keymap <C-c>: Copy file/folder (nvim-tree)
-		vim.keymap.set("n", "<C-c>", api.fs.copy.node, opts("Copy"))
-		-- @keymap <C-v>: Paste file/folder (nvim-tree)
-		vim.keymap.set("n", "<C-v>", api.fs.paste, opts("Paste"))
-		-- @keymap [c: Navigate to previous sibling (nvim-tree)
-		vim.keymap.set("n", "[c", api.node.navigate.sibling.prev, opts("Prev Sibling"))
-		-- @keymap ]c: Navigate to next sibling (nvim-tree)
-		vim.keymap.set("n", "]c", api.node.navigate.sibling.next, opts("Next Sibling"))
-		-- @keymap p: Preview file (nvim-tree)
-		vim.keymap.set("n", "p", api.node.open.preview, opts("Open Preview"))
-		-- @keymap <BS>: Close folder (nvim-tree)
-		vim.keymap.set("n", "<BS>", api.node.navigate.parent_close, opts("Close Directory"))
-		-- @keymap <Tab>: Open preview and keep focus (nvim-tree)
-		vim.keymap.set("n", "<Tab>", api.node.open.preview, opts("Open Preview"))
-		-- @keymap H: Toggle dotfiles (nvim-tree)
-		vim.keymap.set("n", "H", api.tree.toggle_gitignore_filter, opts("Toggle Git Ignore"))
-		-- @keymap I: Toggle dotfiles (nvim-tree)
-		vim.keymap.set("n", "I", api.tree.toggle_hidden_filter, opts("Toggle Dotfiles"))
-	end,
-})
-keymap("n", "<leader>b", ":NvimTreeToggle<CR>", opts)
-
--- ====================================================================================
--- HELP COMMAND
--- ====================================================================================
--- @command Help: Show all keymaps and commands from init.lua
-vim.api.nvim_create_user_command("Help", function()
-	local config_file = vim.fn.stdpath("config") .. "/init.lua"
-	local file = io.open(config_file, "r")
-	if not file then
-		vim.notify("Could not read config file: " .. config_file, vim.log.levels.ERROR)
-		return
-	end
-
-	local keymaps = {}
-	local commands = {}
-	local plugins = {}
-	local current_section = "General"
-
-	for line in file:lines() do
-		-- Parse @keymap comments
-		local key, desc = line:match("^%s*%-%-%s*@keymap%s+(.+):%s*(.+)$")
-		if key and desc then
-			table.insert(keymaps, { key = key, desc = desc, section = current_section })
-		end
-
-		-- Parse @command comments
-		local cmd, cmd_desc = line:match("^%s*%-%-%s*@command%s+(.+):%s*(.+)$")
-		if cmd and cmd_desc then
-			table.insert(commands, { cmd = cmd, desc = cmd_desc })
-		end
-
-		-- Parse @plugin comments
-		local plugin, plugin_desc = line:match("^%s*%-%-%s*@plugin%s+(.+):%s*(.+)$")
-		if plugin and plugin_desc then
-			table.insert(plugins, { plugin = plugin, desc = plugin_desc })
-		end
-
-		-- Track sections
-		if line:match("^%-%-%s*@section") or line:match("^%-%-%s*%-%-%-") then
-			current_section = line:match("UI") and "UI"
-				or line:match("CODING") and "Coding"
-				or line:match("TPOPE") and "Plugins"
-				or line:match("treesitter") and "Treesitter"
-				or "General"
-		end
-	end
-
-	file:close()
-
-	-- Create help buffer
-	local buf = vim.api.nvim_create_buf(false, true)
-	local lines = {}
-
-	table.insert(lines, "Neovim Configuration Help")
-	table.insert(lines, "=" .. string.rep("=", 50))
-	table.insert(lines, "")
-
-	-- Keymaps section
-	table.insert(lines, "KEYMAPS")
-	table.insert(lines, string.rep("-", 50))
-	table.insert(lines, "")
-	for _, km in ipairs(keymaps) do
-		local key = km.key:gsub("<leader>", "<Space>")
-		table.insert(lines, string.format("  %-25s %s", key, km.desc))
-	end
-	table.insert(lines, "")
-
-	-- Commands section
-	if #commands > 0 then
-		table.insert(lines, "COMMANDS")
-		table.insert(lines, string.rep("-", 50))
-		table.insert(lines, "")
-		for _, cmd in ipairs(commands) do
-			table.insert(lines, string.format("  :%-24s %s", cmd.cmd, cmd.desc))
-		end
-		table.insert(lines, "")
-	end
-
-	-- Plugins section
-	if #plugins > 0 then
-		table.insert(lines, "PLUGINS")
-		table.insert(lines, string.rep("-", 50))
-		table.insert(lines, "")
-		for _, plugin in ipairs(plugins) do
-			table.insert(lines, string.format("  %-25s %s", plugin.plugin, plugin.desc))
-		end
-		table.insert(lines, "")
-	end
-
-	table.insert(lines, "=" .. string.rep("=", 50))
-	table.insert(lines, "Press 'q' to close")
-
-	vim.api.nvim_buf_set_lines(buf, 0, -1, false, lines)
-	vim.api.nvim_buf_set_option(buf, "filetype", "help")
-	vim.api.nvim_buf_set_option(buf, "modifiable", false)
-	vim.api.nvim_buf_set_name(buf, "neovim-config-help")
-
-	-- Open in split
-	vim.cmd("split")
-	vim.api.nvim_set_current_buf(buf)
-
-	-- @keymap q: Close help buffer
-	-- Map q to close
-	vim.keymap.set("n", "q", ":close<CR>", { buffer = buf, silent = true })
-end, { desc = "Show all keymaps and commands from init.lua" })
+require("settings")
+require("plugins")
+require("ui")
+require("keymaps")
+require("autocmds")
+require("lsp")
+require("completion")
+require("telescope")
+require("treesitter")
+require("terminal")
+require("utils")
