+++ home-manager/programs/neovim/lua/terminal.lua
@@ -1,0 +1,7 @@
+require('toggleterm').setup({
+  direction = 'float',
+  open_mapping = [[<c-\>]],
+  on_open = function(term)
+    vim.cmd("startinsert!")
+  end
+})
